Expand job list automatically
When an employee is hired, their would-be boss's job is added to the
hire list. Employees can also no longer be hired when there are
insufficient funds to pay their first paycheck. This commit also updates
the "seconds to payday" countdown to match the percentage countdown, as
it was overlooked in the previous commit.