Add a small amount of color to the UI
As a result of this commit, many small improvements were made to the UI
code, especially to its readability and reusability. In addition, the
DeepBinding tool may help to avoid a small amount of boilerplate
property code in the future. Unfortunately, its usage is rare and
complex enough to warrant a reasonable-length Javadoc explanation.